+++ include/delfem2/femnavierstokes.h
@@ -119,9 +119,9 @@ void MergeLinSys_NavierStokes2D(
     const double dt_timestep,
     const double gamma_newmark,
     const double* aXY1,
-    unsigned int nXY,
+    size_t nXY,
     const unsigned int* aTri1,
-    unsigned int nTri,
+    size_t nTri,
     const double* aVal, // vx,vy,press
     const double* aDtVal) // ax,ay,apress
 {
@@ -209,4 +209,4 @@ void MergeLinSys_NavierStokes3D_Dynamic(
 #  include "delfem2/femnavierstokes.cpp"
 #endif
   
-#endif /* fem_ematrix_h */
+#endif /* DFM2_FEMNAVIERSTOKES_H */
